Only raise in production with no email whitelist
See https://github.com/orientation/orientation/issues/205 for the full
conversation about this.

I don't want to cripple (no sign in possible) Orientation by default for this
setting so I want this to return `nil` if not set in every environment except
production.

This means this is now secure by default in production only and elsewhere we
just let anyone in.

It allows me to let anyone sign in to try.orientation.io