Only auto-sign in with the first user in development
This was a very bad idea to begin with but the fact that I added this code
made it much worse:

    (!Rails.env.production? or !Rails.env.staging?)

This always returned true in production, meaning that authentication was disabled
because of the `or !Rails.env.staging?`.

This commit will break try.orientation.io but that's okay because I was lazy
and should have found an alternative way to bypass authentication for it.

In summary if you're currently running a version of Orientation with this
commit as the latest change to application_controller.rb, you're Orientation
installation is wide open because authentication is bypassed:
https://github.com/orientation/orientation/commit/e71754ef18c7f2d7a202d3b2eecc5ef1f9478b9c

Please update as soon as possible, shame me, and make me feel bad for not writing
better test for the authentication flow.