Escape all HTML input from article content
This will likely affect users who relied on HTML tags and JavaScript embed
in their article content but it's safer for everyone to default to not allowing
any HTML content to be evaluated.

It would be a stored XSS attack vector if someone somehow managed to
alter an article. They could steal another user's session to impersonate them.

Many thanks to @nsheridan for raising this issue in #217.

Fixes #217.